feat(collective): widen comm_id to u64 and harden split correctness
Expand communicator ID from u32 to u64 throughout the wire protocol,
router, and split logic. The u64 FNV-1a hash now produces the full
64-bit output directly, removing the prior fold-to-u32 step that
introduced unnecessary collision risk.

Split operation changes:
- Exchange includes a split_generation counter (16-byte payload up from 8)
- Validate that all ranks share the same generation value before forming
  the new communicator; mismatches surface as a CollectiveFailed error
  rather than silently producing inconsistent groups
- config is propagated into child communicators via Arc::clone

Collective correctness improvements across allreduce, allgather,
alltoall, reduce_scatter, scan, barrier, and pipelined allreduce:
- Replace tokio::join! + paired ? with tokio::try_join! so the first
  error immediately cancels the peer future rather than ignoring it

Additional changes:
- Add byte_transport module for raw byte send/recv (extracted from
  async_client.rs to keep file size manageable)
- Add module-level documentation to compression/mod.rs covering when to
  use each compressor and the error-feedback residual pattern
- Barrier and BarrierAck messages updated to carry comm_id: u64

+++ nexar/src/protocol/message.rs
@@ -31,10 +31,10 @@ pub enum NexarMessage {
     },
 
     /// Barrier request: all ranks must reach this epoch before proceeding.
-    Barrier { epoch: u64, comm_id: u32 },
+    Barrier { epoch: u64, comm_id: u64 },
 
     /// Barrier acknowledgement from coordinator.
-    BarrierAck { epoch: u64, comm_id: u32 },
+    BarrierAck { epoch: u64, comm_id: u64 },
 
     /// Periodic heartbeat for failure detection.
     Heartbeat { timestamp_ns: u64 },
